double data type

+++ main.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    int a, b;
+    double a, b;
 
     cout << "Enter first number: ";
     cin >> a;
@@ -11,7 +11,7 @@ int main() {
     cin >> b;
 
     cout << "Sum of " << a << " and " << b << "= "<<a+b<< endl;
-    
+
 
     return 0;
 }
